Fix: CorsConfig y exposición de información sensible
