+++ .gitignore
@@ -31,3 +31,7 @@ build/
 
 ### VS Code ###
 .vscode/
+
+### Spring Boot Configuration Files with Sensitive Data ###
+application.properties
+*.env
