+++ src/main/java/com/example/financelloapi/config/CorsConfig.java
@@ -1,10 +1,34 @@
 package com.example.financelloapi.config;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.List;
 
 @Configuration
 public class CorsConfig {
-    // CORS configuration is now handled by SecurityConfig to avoid conflicts
-    // The SecurityConfig.corsConfigurationSource() method manages all CORS settings
+
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        CorsConfiguration config = new CorsConfiguration();
+        
+        // Permitir múltiples orígenes: localhost para desarrollo y Netlify para producción
+        config.setAllowedOriginPatterns(List.of(
+            "http://localhost:4200",
+            "https://685c8b9d9b4f1900080580ee--financello-web-v1.netlify.app"
+        ));
+        
+        config.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS"));
+        config.setAllowedHeaders(List.of("*"));
+        config.setAllowCredentials(true);
+        config.setExposedHeaders(List.of("Authorization", "Content-Type"));
+
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", config);
+        return source;
+    }
 }
